assignment 3

+++ app/Http/Controllers/CarController.php
@@ -37,8 +37,8 @@ class CarController extends Controller
      */
     public function store(Car $car , Request $request)
     {
-        Cars::create($this->validateCar($request));
-        return redirect(route('cars.index'));
+        Car::create($this->validateCar($request));
+        return redirect('/cars');
     }
 
     /**
@@ -49,7 +49,9 @@ class CarController extends Controller
      */
     public function show(Car $car)
     {
-        //
+
+        $cars = Car::orderBy('registration_number', 'asc')->get();
+        return view('cars.index', compact('cars'));
     }
 
     /**
@@ -60,6 +62,7 @@ class CarController extends Controller
      */
     public function edit(Car $car)
     {
+        $this->authorize('update-car');
         return view('cars.edit',compact('car'));
     }
 
@@ -72,6 +75,7 @@ class CarController extends Controller
      */
     public function update(Request $request, Car $car)
     {
+        $this->authorize('update-car');
         $car->update($this->validateCar($request));
         return redirect(route('cars.show' , $car));
     }
@@ -84,6 +88,7 @@ class CarController extends Controller
      */
     public function destroy(Car $car)
     {
+        $this->authorize('update-car');
         $car->delete();
         return redirect(route('cars.index'));
     }
@@ -91,9 +96,9 @@ class CarController extends Controller
     public function validateCar($request){
 
         return $request->validate([
-            'registration_number'=>'required|numeric|unique',
+            'registration_number'=>'required|numeric',
             'manufacturer'=>'required',
-            'currently_available'=>'boolean'
+            'currently_available'=>'required'
         ]);
     }
 }
